+++ include/boost/callable_traits/detail/unguarded/function.hpp
@@ -11,7 +11,6 @@ DO NOT INCLUDE THIS HEADER DIRECTLY
 #define BOOST_CLBL_TRTS_INCLUDE_TRANSACTION_SAFE
 #define BOOST_CLBL_TRTS_IS_TRANSACTION_SAFE std::false_type
 #include <boost/callable_traits/detail/unguarded/function_2.hpp>
-
 #undef BOOST_CLBL_TRTS_INCLUDE_TRANSACTION_SAFE
 #undef BOOST_CLBL_TRTS_IS_TRANSACTION_SAFE
 
@@ -19,7 +18,6 @@ DO NOT INCLUDE THIS HEADER DIRECTLY
 #define BOOST_CLBL_TRTS_IS_TRANSACTION_SAFE std::true_type
 #define BOOST_CLBL_TRTS_INCLUDE_TRANSACTION_SAFE transaction_safe
 #include <boost/callable_traits/detail/unguarded/function_2.hpp>
-#endif
-
 #undef BOOST_CLBL_TRTS_INCLUDE_TRANSACTION_SAFE
 #undef BOOST_CLBL_TRTS_IS_TRANSACTION_SAFE
+#endif // #ifdef BOOST_CLBL_TRTS_ENABLE_TRANSACTION_SAFE
