+++ include/boost/callable_traits/detail/unguarded/function_2.hpp
@@ -6,254 +6,18 @@ Distributed under the Boost Software License, Version 1.0.
 
 HEADER GUARDS INTENTIONALLY OMITTED
 DO NOT INCLUDE THIS HEADER DIRECTLY
-
-macros used:
-
-BOOST_CLBL_TRTS_INCLUDE_QUALIFIERS - the function-level qualifiers for the
-    current inclusion (combinations of `const` `volatile` `&` `&&`, or nothing)
-
-BOOST_CLBL_TRTS_INCLUDE_TRANSACTION_SAFE - the transaction_safe specifier for
-    the current include (`transaction_safe` or nothing)
-
-BOOST_CLBL_TRTS_IS_TRANSACTION_SAFE - `std::true_type` or `std::false_type`,
-    tied on whether BOOST_CLBL_TRTS_INCLUDE_TRANSACTION_SAFE is `transaction_safe`
-
-BOOST_CLBL_TRTS_TRANSACTION_SAFE_SPECIFIER - `transaction_safe` when
-    BOOST_CLBL_TRTS_ENABLE_TRANSACTION_SAFE is enabled, otherwise nothing
-
-BOOST_CLBL_TRTS_NOEXCEPT_SPEC - the noexcept specifier for
-    the current include (`noexcept` or nothing)
-
-BOOST_CLBL_TRTS_IS_NOEXCEPT - `std::true_type` or `std::false_type`,
-    tied on whether BOOST_CLBL_TRTS_NOEXCEPT_SPEC is `noexcept`
-
-BOOST_CLBL_TRTS_NOEXCEPT_SPECIFIER - `noexcept` if
-    BOOST_CLBL_TRTS_ENABLE_NOEXCEPT_TYPES is defined, otherwise nothing
-
 */
 
+#define BOOST_CLBL_TRTS_NOEXCEPT_SPEC
+#define BOOST_CLBL_TRTS_IS_NOEXCEPT std::false_type
+#include <boost/callable_traits/detail/unguarded/function_3.hpp>
+#undef BOOST_CLBL_TRTS_INCLUDE_NOEXCEPT
+#undef BOOST_CLBL_TRTS_IS_NOEXCEPT
 
-
-template<typename Return, typename... Args BOOST_CLBL_TRTS_NOEXCEPT_TEMPL>
-struct function<Return(Args...)
-    BOOST_CLBL_TRTS_INCLUDE_QUALIFIERS
-    BOOST_CLBL_TRTS_INCLUDE_TRANSACTION_SAFE
-    BOOST_CLBL_TRTS_NOEXCEPT_SPEC>
- : default_callable_traits<dummy BOOST_CLBL_TRTS_INCLUDE_QUALIFIERS> {
-     
-    static constexpr bool value = true;
-    
-    using traits = function;
-
-    using return_type = Return;
-
-    using arg_types = std::tuple<Args...>;
-    using non_invoke_arg_types = arg_types;
-
-    using type = Return(Args...)
-        BOOST_CLBL_TRTS_INCLUDE_QUALIFIERS
-        BOOST_CLBL_TRTS_INCLUDE_TRANSACTION_SAFE
-        BOOST_CLBL_TRTS_NOEXCEPT_SPEC;
-
-    using function_type = Return(Args...);
-
-    using qualified_function_type = Return(Args...)
-        BOOST_CLBL_TRTS_INCLUDE_QUALIFIERS
-        BOOST_CLBL_TRTS_INCLUDE_TRANSACTION_SAFE
-        BOOST_CLBL_TRTS_NOEXCEPT_SPEC;
-
-    using remove_varargs = type;
-
-    using add_varargs = Return (Args..., ...)
-        BOOST_CLBL_TRTS_INCLUDE_QUALIFIERS
-        BOOST_CLBL_TRTS_INCLUDE_TRANSACTION_SAFE
-        BOOST_CLBL_TRTS_NOEXCEPT_SPEC;
-    
-    using is_noexcept = BOOST_CLBL_TRTS_IS_NOEXCEPT;
-
-    using remove_noexcept = Return(Args...)
-        BOOST_CLBL_TRTS_INCLUDE_QUALIFIERS
-        BOOST_CLBL_TRTS_INCLUDE_TRANSACTION_SAFE;
-
-    using add_noexcept = Return(Args...)
-        BOOST_CLBL_TRTS_INCLUDE_QUALIFIERS
-        BOOST_CLBL_TRTS_INCLUDE_TRANSACTION_SAFE
-        BOOST_CLBL_TRTS_NOEXCEPT_SPECIFIER;
-
-    using is_transaction_safe = BOOST_CLBL_TRTS_IS_TRANSACTION_SAFE;
-
-    using remove_transaction_safe = Return(Args...)
-        BOOST_CLBL_TRTS_INCLUDE_QUALIFIERS
-        BOOST_CLBL_TRTS_NOEXCEPT_SPEC;
-
-    using add_transaction_safe = Return(Args...)
-        BOOST_CLBL_TRTS_INCLUDE_QUALIFIERS
-        BOOST_CLBL_TRTS_TRANSACTION_SAFE_SPECIFIER
-        BOOST_CLBL_TRTS_NOEXCEPT_SPEC;
-
-    using qualifiers = default_callable_traits<dummy BOOST_CLBL_TRTS_INCLUDE_QUALIFIERS>;
-    
-    template<qualifier_flags Flags>
-    using set_qualifiers = set_function_qualifiers<Flags, is_transaction_safe::value,
-        is_noexcept::value, Return, Args...>;
-    
-    #ifdef BOOST_CLBL_TRTS_DISABLE_ABOMINABLE_FUNCTIONS
-
-    using add_member_lvalue_reference = abominable_functions_not_supported_on_this_compiler;
-    using add_member_rvalue_reference = abominable_functions_not_supported_on_this_compiler;
-    using add_member_const = abominable_functions_not_supported_on_this_compiler;
-    using add_member_volatile = abominable_functions_not_supported_on_this_compiler;
-    using add_member_cv = abominable_functions_not_supported_on_this_compiler;
-
-    #else
-
-    using add_member_lvalue_reference = set_qualifiers<
-        collapse_flags<qualifiers::q_flags, lref_>::value>;
-
-    using add_member_rvalue_reference = set_qualifiers<
-        collapse_flags<qualifiers::q_flags, rref_>::value>;
-
-    using add_member_const = set_qualifiers<qualifiers::q_flags | const_>;
-
-    using add_member_volatile = set_qualifiers<qualifiers::q_flags | volatile_>;
-
-    using add_member_cv = set_qualifiers<qualifiers::q_flags | cv_>;
-
-    #endif // #ifdef BOOST_CLBL_TRTS_DISABLE_ABOMINABLE_FUNCTIONS
-    
-    using remove_member_reference = set_qualifiers<qualifiers::cv_flags>;
-
-    using remove_member_const = set_qualifiers<
-        qualifiers::ref_flags | remove_const_flag<qualifiers::cv_flags>::value>;
-        
-    using remove_member_volatile = set_qualifiers<
-        qualifiers::ref_flags | remove_volatile_flag<qualifiers::cv_flags>::value>;
-        
-    using remove_member_cv = set_qualifiers<qualifiers::ref_flags>;
-
-    template<typename U>
-    using apply_member_pointer = add_member_pointer<type, U>;
-    
-    template<typename NewReturn>
-    using apply_return = NewReturn(Args...)
-        BOOST_CLBL_TRTS_INCLUDE_QUALIFIERS
-        BOOST_CLBL_TRTS_INCLUDE_TRANSACTION_SAFE
-        BOOST_CLBL_TRTS_NOEXCEPT_SPEC;
-    
-    template<template<class...> class Container>
-    using expand_args = Container<Args...>;
-};
-
-
-template<typename Return, typename... Args BOOST_CLBL_TRTS_NOEXCEPT_TEMPL>
-struct function<Return (Args..., ...)
-    BOOST_CLBL_TRTS_INCLUDE_QUALIFIERS
-    BOOST_CLBL_TRTS_INCLUDE_TRANSACTION_SAFE
-    BOOST_CLBL_TRTS_NOEXCEPT_SPEC>
- : default_callable_traits<> {
-     
-    static constexpr bool value = true;
-    
-    using has_varargs = std::true_type;
-    using traits = function;
-    using return_type = Return;
-    using arg_types = std::tuple<Args...>;
-
-    using type = Return (Args..., ...)
-        BOOST_CLBL_TRTS_INCLUDE_QUALIFIERS
-        BOOST_CLBL_TRTS_INCLUDE_TRANSACTION_SAFE
-        BOOST_CLBL_TRTS_NOEXCEPT_SPEC;
-
-    using function_type = Return(Args..., ...);
-
-    using qualified_function_type = Return(Args..., ...)
-        BOOST_CLBL_TRTS_INCLUDE_QUALIFIERS
-        BOOST_CLBL_TRTS_INCLUDE_TRANSACTION_SAFE
-        BOOST_CLBL_TRTS_NOEXCEPT_SPEC;
-
-    using remove_varargs = Return (Args...)
-        BOOST_CLBL_TRTS_INCLUDE_QUALIFIERS
-        BOOST_CLBL_TRTS_INCLUDE_TRANSACTION_SAFE
-        BOOST_CLBL_TRTS_NOEXCEPT_SPEC;
-
-    using add_varargs = type;
-
-    using is_noexcept = BOOST_CLBL_TRTS_IS_NOEXCEPT;
-
-    using remove_noexcept = Return(Args..., ...)
-        BOOST_CLBL_TRTS_INCLUDE_QUALIFIERS
-        BOOST_CLBL_TRTS_INCLUDE_TRANSACTION_SAFE;
-
-    using add_noexcept = Return(Args..., ...)
-        BOOST_CLBL_TRTS_INCLUDE_QUALIFIERS
-        BOOST_CLBL_TRTS_INCLUDE_TRANSACTION_SAFE
-        BOOST_CLBL_TRTS_NOEXCEPT_SPECIFIER;
-
-    using is_transaction_safe = BOOST_CLBL_TRTS_IS_TRANSACTION_SAFE;
-
-    using remove_transaction_safe = Return(Args..., ...)
-        BOOST_CLBL_TRTS_INCLUDE_QUALIFIERS
-        BOOST_CLBL_TRTS_NOEXCEPT_SPEC;
-
-    using add_transaction_safe = Return(Args..., ...)
-        BOOST_CLBL_TRTS_INCLUDE_QUALIFIERS
-        BOOST_CLBL_TRTS_TRANSACTION_SAFE_SPECIFIER
-        BOOST_CLBL_TRTS_NOEXCEPT_SPEC;
-
-    using qualifiers = default_callable_traits<dummy BOOST_CLBL_TRTS_INCLUDE_QUALIFIERS>;
-    
-    template<qualifier_flags Flags>
-    using set_qualifiers = set_varargs_function_qualifiers<Flags, is_transaction_safe::value,
-        is_noexcept::value, Return, Args...>;
-    
-    #ifdef BOOST_CLBL_TRTS_DISABLE_ABOMINABLE_FUNCTIONS
-
-    using add_member_lvalue_reference = abominable_functions_not_supported_on_this_compiler;
-    using add_member_rvalue_reference = abominable_functions_not_supported_on_this_compiler;
-    using add_member_const = abominable_functions_not_supported_on_this_compiler;
-    using add_member_volatile = abominable_functions_not_supported_on_this_compiler;
-    using add_member_cv = abominable_functions_not_supported_on_this_compiler;
-
-    #else
-
-    using add_member_lvalue_reference = set_qualifiers<
-        collapse_flags<qualifiers::q_flags, lref_>::value>;
-        
-    using add_member_rvalue_reference = set_qualifiers<
-        collapse_flags<qualifiers::q_flags, rref_>::value>;
-        
-    using add_member_const = set_qualifiers<qualifiers::q_flags | const_>;
-
-    using add_member_volatile = set_qualifiers<qualifiers::q_flags | volatile_>;
-
-    using add_member_cv = set_qualifiers<qualifiers::q_flags | cv_>;
-    
-    #endif // #ifdef BOOST_CLBL_TRTS_DISABLE_ABOMINABLE_FUNCTIONS
-
-    using remove_member_reference = set_qualifiers<qualifiers::cv_flags>;
-
-    using remove_member_const = set_qualifiers<
-        qualifiers::ref_flags | remove_const_flag<qualifiers::cv_flags>::value>;
-        
-    using remove_member_volatile = set_qualifiers<
-        qualifiers::ref_flags | remove_volatile_flag<qualifiers::cv_flags>::value>;
-        
-    using remove_member_cv = set_qualifiers<qualifiers::ref_flags>;
-    
-    template<typename U>
-    using apply_member_pointer =
-        Return( BOOST_CLBL_TRTS_DEFAULT_VARARGS_CC U::*)(Args..., ...)
-            BOOST_CLBL_TRTS_INCLUDE_QUALIFIERS
-            BOOST_CLBL_TRTS_INCLUDE_TRANSACTION_SAFE
-            BOOST_CLBL_TRTS_NOEXCEPT_SPEC;
-        
-    template<typename NewReturn>
-    using apply_return = NewReturn(Args..., ...)
-        BOOST_CLBL_TRTS_INCLUDE_QUALIFIERS
-        BOOST_CLBL_TRTS_INCLUDE_TRANSACTION_SAFE
-        BOOST_CLBL_TRTS_NOEXCEPT_SPEC;
-    
-    template<template<class...> class Container>
-    using expand_args = Container<Args...>;
-
-};
+#ifdef BOOST_CLBL_TRTS_ENABLE_NOEXCEPT_TYPES
+#define BOOST_CLBL_TRTS_NOEXCEPT_SPEC noexcept
+#define BOOST_CLBL_TRTS_IS_NOEXCEPT std::true_type
+#include <boost/callable_traits/detail/unguarded/function_3.hpp>
+#undef BOOST_CLBL_TRTS_INCLUDE_NOEXCEPT
+#undef BOOST_CLBL_TRTS_IS_NOEXCEPT
+#endif // #ifdef BOOST_CLBL_TRTS_ENABLE_NOEXCEPT_TYPES
